struct(T3): rename game.rs to core.rs and replace re-export facade with explicit imports
