+++ src/game/events_apply.rs
@@ -1,6 +1,8 @@
 //! Event outcomes: applies random and historical events to the game state.
 //! Trigger selection lives in `events.rs`; week orchestration stays in `turn.rs`.
 
+use rand::Rng;
+
 use super::constants;
 use super::*;
