+++ src/game/mod.rs
@@ -6,11 +6,10 @@
 mod actions;
 pub mod band;
 mod constants;
+pub mod core;
 mod economy;
 pub mod events;
 mod events_apply;
-#[allow(clippy::module_inception)]
-pub mod game;
 pub mod genre;
 pub mod music;
 pub mod player;
@@ -24,26 +23,4 @@ pub mod world;
 mod tests;
 
 pub use constants::{BREAK_WEEKS, PRESSING_TIERS};
-pub use game::{Game, GameAction, SupportTourOffer};
-
-// Re-exports for submodules and external crates
-#[allow(unused_imports)]
-pub use crate::data_loader::GameDataFiles;
-#[allow(unused_imports)]
-pub use crate::game::music::{
-    ActiveMarketingCampaign, MarketingCampaignType, Release, ReleaseType,
-};
-#[allow(unused_imports)]
-pub use band::Band;
-#[allow(unused_imports)]
-pub use events::EventManager;
-#[allow(unused_imports)]
-pub use player::Player;
-#[allow(unused_imports)]
-pub use rand::rngs::StdRng;
-#[allow(unused_imports)]
-pub use rand::{Rng, SeedableRng};
-#[allow(unused_imports)]
-pub use timeline::MusicTimeline;
-#[allow(unused_imports)]
-pub use world::{GameWorld, PotentialDealOffer};
+pub use core::{Game, GameAction, SupportTourOffer};
